api/middleware/request_id.py
from __future__ import annotations

import uuid

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.requests import Request
from starlette.responses import Response
from starlette.types import ASGIApp


class RequestIDMiddleware(BaseHTTPMiddleware):
    """
    Propagate or generate a unique request ID for every request.

    - Reads `X-Request-ID` from the incoming request headers if present.
    - Falls back to a freshly generated UUID4.
    - Attaches the ID to `request.state.request_id`.
    - Adds `X-Request-ID` to the outgoing response headers.
    """

    def __init__(self, app: ASGIApp) -> None:
        super().__init__(app)

    async def dispatch(self, request: Request, call_next) -> Response:
        request_id = request.headers.get("X-Request-ID") or str(uuid.uuid4())
        request.state.request_id = request_id

        response = await call_next(request)
        response.headers["X-Request-ID"] = request_id
        return response
