requirements_api.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
alembic>=1.13
asyncpg>=0.29
pydantic-settings>=2.2
arq>=0.25
redis>=5.0
httpx>=0.27
google-auth>=2.29
authlib>=1.3
python-jose[cryptography]>=3.3
cryptography>=42.0
python-multipart>=0.0.9
aioboto3>=12.3
markdown>=3.6
user-agents>=2.2.0
sentry-sdk>=2.0
opentelemetry-api>=1.24
opentelemetry-sdk>=1.24
pytest-asyncio>=0.23
anyio>=4.3
